move `JavaCodeUnit.Parameter` to toplevel `JavaParameter`
Originally I decided for `JavaCodeUnit.Parameter` because
* the concept is very closely related to `JavaCodeUnit`
* as a nested class it would enable to import as `Parameter` in a clear narrow scope or as `JavaCodeUnit.Parameter` in a wider scope
* I considered "parameter" alone to be too generic as a name

However, I have now decided to name it `JavaParameter` after all. The reason is simply, that it is consistent with the Reflection API. And since we followed this philosophy with most other classes (`Class` -> `JavaClass`, `Method` -> `JavaMethod`, ...) I consider this to be the way of least surprise. Which in the end is one of the most important guiding principles we should follow.
